+++ src/ndv/views/bases/__init__.py
@@ -1,9 +1,9 @@
 from ._array_view import ArrayView
 from ._lut_view import LutView
-from ._mouseable import Mouseable, filter_mouse_events
-from .graphics._canvas import ArrayCanvas
+from ._view_base import Viewable
+from .graphics._canvas import ArrayCanvas, HistogramCanvas
 from .graphics._canvas_elements import CanvasElement, ImageHandle, RoiHandle
-from .graphics._histogram import HistogramCanvas
+from .graphics._mouseable import Mouseable, filter_mouse_events
 
 __all__ = [
     "ArrayCanvas",
@@ -14,5 +14,6 @@ __all__ = [
     "LutView",
     "Mouseable",
     "RoiHandle",
+    "Viewable",
     "filter_mouse_events",
 ]
